refactor:重新组织store目录结构
