+++ frontend/src/components/SettingsPanel.tsx
@@ -1,7 +1,6 @@
 import { Settings as SettingsIcon, Cpu, Database, Network, Activity, Save, RefreshCw, Undo2, TerminalSquare, Search, Eye, Sparkles } from 'lucide-react';
 import { useSettings } from '../hooks/useSettings';
-import { useToastStore } from '@/stores/toastStore';
-import { useChatUiStore } from '@/stores/chatUiStore';
+import { useChatUiStore, useToastStore } from '@/stores';
 import { GeneralSettings } from './settings/GeneralSettings';
 import { ModelSettings } from './settings/ModelSettings';
 import { InfrastructureSettings } from './settings/InfrastructureSettings';
