+++ frontend/src/stores/kernel/index.ts
@@ -1,0 +1,1 @@
+export { useKernelStore } from './kernelStore';
