+++ frontend/src/stores/kernelStore.ts
@@ -1,936 +1,1 @@
-/**
- * Kernel Store - Zustand state management for kernel terminal
- *
- * Features:
- * - WebSocket connection to backend log streaming
- * - BroadcastChannel for cross-window synchronization
- * - Primary/secondary window election
- * - Automatic reconnection with exponential backoff
- * - Log buffering and filtering
- */
-
-import { create } from 'zustand';
-import { devtools, persist } from 'zustand/middleware';
-import { useMemoryTaskStore } from '@/stores/memoryTaskStore';
-import type {
-  KernelStore,
-  LogEntry,
-  BroadcastMessage,
-  TraceGroup,
-  RuntimeEvent,
-} from '@/types/kernel';
-
-// Configuration constants
-const BACKEND_ORIGIN = import.meta.env.VITE_BACKEND_ORIGIN || window.location.origin;
-const API_ORIGIN = new URL(BACKEND_ORIGIN, window.location.origin);
-const WS_PROTOCOL = API_ORIGIN.protocol === 'https:' ? 'wss' : 'ws';
-const WS_URL = `${WS_PROTOCOL}://${API_ORIGIN.host}/api/v1/ws/logs`;
-const RUNTIME_EVENTS_URL = `${API_ORIGIN.origin}/api/v1/runtime-events/stream`;
-
-const BROADCAST_CHANNEL_NAME = 'hivememory_kernel_logs';
-const PRIMARY_WINDOW_KEY = 'hivememory_primary_window';
-const HEARTBEAT_INTERVAL = 5000; // 5 seconds
-const MAX_RECONNECT_ATTEMPTS = 10;
-const RECONNECT_DELAYS = [1000, 2000, 4000, 8000, 16000, 30000]; // Exponential backoff
-const PING_INTERVAL = 30000; // 30 seconds
-const CONNECT_TIMEOUT = 10000;
-type PingWebSocket = WebSocket & { _pingTimer?: ReturnType<typeof setInterval> };
-
-/**
- * Group logs by trace_id and span_name
- */
-function groupLogsByTrace(logs: LogEntry[]): Map<string, TraceGroup> {
-  const groups = new Map<string, TraceGroup>();
-
-  for (const log of logs) {
-    const { trace_id, span_name, task_type } = log;
-
-    if (!groups.has(trace_id)) {
-      groups.set(trace_id, {
-        trace_id,
-        spans: new Map(),
-        task_type,
-        collapsed: task_type === 'background',
-      });
-    }
-
-    const trace = groups.get(trace_id)!;
-    if (!trace.spans.has(span_name)) {
-      trace.spans.set(span_name, {
-        span_name,
-        logs: [],
-        collapsed: false,
-        task_type,
-        first_timestamp: log.timestamp,
-        last_timestamp: log.timestamp,
-      });
-    }
-
-    const span = trace.spans.get(span_name)!;
-    span.logs.push(log);
-    span.last_timestamp = log.timestamp;
-  }
-
-  return groups;
-}
-
-/**
- * Check if this window should be the primary window
- * Primary window maintains the WebSocket connection
- */
-function checkAndClaimPrimaryWindow(): boolean {
-  const now = Date.now();
-  const stored = localStorage.getItem(PRIMARY_WINDOW_KEY);
-
-  if (!stored) {
-    // No primary window exists, claim it
-    localStorage.setItem(PRIMARY_WINDOW_KEY, now.toString());
-    return true;
-  }
-
-  const lastHeartbeat = parseInt(stored, 10);
-  if (now - lastHeartbeat > HEARTBEAT_INTERVAL * 2) {
-    // Primary window is dead, take over
-    localStorage.setItem(PRIMARY_WINDOW_KEY, now.toString());
-    return true;
-  }
-
-  return false;
-}
-
-/**
- * Initialize WebSocket connection (primary window only)
- */
-function initializeWebSocket(
-  set: (partial: Partial<KernelStore>) => void,
-  get: () => KernelStore
-) {
-  let ws: WebSocket;
-  try {
-    ws = new WebSocket(WS_URL);
-  } catch (error) {
-    const message =
-      error instanceof Error ? error.message : 'WebSocket initialization failed';
-    console.error('[KernelStore] Failed to initialize WebSocket:', error);
-    set({
-      connection: {
-        ...get().connection,
-        status: 'error',
-        error: message,
-      },
-    });
-    return;
-  }
-
-  const connectTimeout = setTimeout(() => {
-    if (ws.readyState === WebSocket.CONNECTING) {
-      console.error('[KernelStore] WebSocket connection timeout');
-      set({
-        connection: {
-          ...get().connection,
-          status: 'error',
-          error: 'WebSocket connection timeout',
-        },
-      });
-      ws.close();
-    }
-  }, CONNECT_TIMEOUT);
-
-  ws.onopen = () => {
-    const managedWs = ws as PingWebSocket;
-    console.log('[KernelStore] WebSocket connected');
-    clearTimeout(connectTimeout);
-    set({
-      connection: {
-        status: 'connected',
-        error: null,
-        connectedAt: Date.now(),
-        reconnectAttempts: 0,
-        lastPingTime: Date.now(),
-      },
-    });
-
-    // Start ping interval
-    const pingTimer = setInterval(() => {
-      if (ws.readyState === WebSocket.OPEN) {
-        ws.send('ping');
-        set({
-          connection: {
-            ...get().connection,
-            lastPingTime: Date.now(),
-          },
-        });
-      }
-    }, PING_INTERVAL);
-
-    // Store ping timer for cleanup
-    managedWs._pingTimer = pingTimer;
-  };
-
-  ws.onmessage = (event) => {
-    try {
-      // Handle pong response
-      if (event.data === 'pong') {
-        return;
-      }
-
-      const logData = JSON.parse(event.data);
-
-      // Validate log structure
-      if (!logData.timestamp || !logData.level || !logData.logger) {
-        console.warn('[KernelStore] Invalid log entry:', logData);
-        return;
-      }
-
-      // Add log to store
-      get().addLog(logData);
-
-      // Broadcast to other windows
-      const state = get();
-      if (state._isPrimaryWindow && state._broadcastChannel) {
-        state._broadcastChannel.postMessage({
-          type: 'NEW_LOG',
-          payload: logData,
-        });
-      }
-    } catch (error) {
-      console.error('[KernelStore] Failed to parse WebSocket message:', error);
-    }
-  };
-
-  ws.onerror = (error) => {
-    console.error('[KernelStore] WebSocket error:', error);
-    clearTimeout(connectTimeout);
-    set({
-      connection: {
-        ...get().connection,
-        status: 'error',
-        error: 'WebSocket connection error. Check backend status or browser security policy.',
-      },
-    });
-  };
-
-  ws.onclose = () => {
-    const managedWs = ws as PingWebSocket;
-    console.log('[KernelStore] WebSocket closed');
-    clearTimeout(connectTimeout);
-
-    // Clear ping timer
-    if (managedWs._pingTimer) {
-      clearInterval(managedWs._pingTimer);
-    }
-
-    const state = get();
-    if (state._manualDisconnecting) {
-      set({
-        connection: {
-          status: 'disconnected',
-          error: null,
-          connectedAt: null,
-          reconnectAttempts: 0,
-          lastPingTime: null,
-        },
-        _manualDisconnecting: false,
-        _ws: null,
-        _reconnectTimer: null,
-      });
-      return;
-    }
-
-    const attempts = state.connection.reconnectAttempts;
-
-    if (attempts < MAX_RECONNECT_ATTEMPTS) {
-      // Schedule reconnection
-      const delay =
-        RECONNECT_DELAYS[Math.min(attempts, RECONNECT_DELAYS.length - 1)];
-
-      console.log(
-        `[KernelStore] Reconnecting in ${delay}ms (attempt ${attempts + 1}/${MAX_RECONNECT_ATTEMPTS})`
-      );
-
-      set({
-        connection: {
-          ...state.connection,
-          status: 'reconnecting',
-          reconnectAttempts: attempts + 1,
-        },
-      });
-
-      const timer = setTimeout(() => {
-        if (get()._isPrimaryWindow) {
-          initializeWebSocket(set, get);
-        }
-      }, delay);
-
-      set({ _reconnectTimer: timer });
-    } else {
-      set({
-        connection: {
-          ...state.connection,
-          status: 'error',
-          error: 'Max reconnection attempts reached',
-        },
-      });
-    }
-  };
-
-  set({ _ws: ws });
-}
-
-function initializeRuntimeEventStream(
-  set: (partial: Partial<KernelStore>) => void,
-  get: () => KernelStore
-) {
-  const state = get();
-  if (
-    state._eventSource &&
-    state.runtimeEventConnection.status !== 'disabled' &&
-    state.runtimeEventConnection.status !== 'error'
-  ) {
-    return;
-  }
-
-  const lastEventId = state.runtimeEventConnection.lastEventId;
-  const url = new URL(RUNTIME_EVENTS_URL);
-  if (lastEventId) {
-    url.searchParams.set('last_event_id', lastEventId);
-  }
-
-  set({
-    runtimeEventConnection: {
-      ...state.runtimeEventConnection,
-      status: 'connecting',
-      error: null,
-    },
-  });
-
-  void fetch(`${API_ORIGIN.origin}/api/v1/runtime-events/status`)
-    .then((response) => response.ok ? response.json() : { enabled: false })
-    .then((status: { enabled?: boolean }) => {
-      if (!status.enabled) {
-        set({
-          runtimeEventConnection: {
-            ...get().runtimeEventConnection,
-            status: 'disabled',
-            error: 'RuntimeEvent stream disabled',
-          },
-        });
-        return false;
-      }
-      return true;
-    })
-    .then((enabled) => {
-      if (!enabled) return;
-      openRuntimeEventSource(url, set, get);
-    })
-    .catch((error) => {
-      const message =
-        error instanceof Error ? error.message : 'RuntimeEvent status check failed';
-      set({
-        runtimeEventConnection: {
-          ...get().runtimeEventConnection,
-          status: 'error',
-          error: message,
-        },
-      });
-    });
-}
-
-function openRuntimeEventSource(
-  url: URL,
-  set: (partial: Partial<KernelStore>) => void,
-  get: () => KernelStore
-) {
-  try {
-    const eventSource = new EventSource(url.toString());
-
-    eventSource.onopen = () => {
-      set({
-        runtimeEventConnection: {
-          ...get().runtimeEventConnection,
-          status: 'connected',
-          connectedAt: Date.now(),
-          error: null,
-        },
-      });
-    };
-
-    eventSource.addEventListener('runtime_event', (message) => {
-      try {
-        const runtimeEvent = JSON.parse(message.data) as RuntimeEvent;
-        get().addRuntimeEvent(runtimeEvent);
-
-        const current = get();
-        if (current._isPrimaryWindow && current._broadcastChannel) {
-          current._broadcastChannel.postMessage({
-            type: 'NEW_RUNTIME_EVENT',
-            payload: runtimeEvent,
-          });
-        }
-      } catch (error) {
-        console.error('[KernelStore] Failed to parse RuntimeEvent:', error);
-      }
-    });
-
-    eventSource.onerror = () => {
-      const currentSource = get()._eventSource;
-      const status = currentSource?.readyState === EventSource.CLOSED ? 'disabled' : 'error';
-      set({
-        runtimeEventConnection: {
-          ...get().runtimeEventConnection,
-          status,
-          error:
-            status === 'disabled'
-              ? 'RuntimeEvent stream disabled'
-              : 'RuntimeEvent stream connection error',
-        },
-      });
-    };
-
-    set({ _eventSource: eventSource });
-  } catch (error) {
-    const message =
-      error instanceof Error ? error.message : 'RuntimeEvent stream initialization failed';
-    set({
-      runtimeEventConnection: {
-        ...get().runtimeEventConnection,
-        status: 'error',
-        error: message,
-      },
-    });
-  }
-}
-
-/**
- * Handle BroadcastChannel messages (all windows)
- */
-function handleBroadcastMessage(
-  message: BroadcastMessage,
-  set: (partial: Partial<KernelStore>) => void,
-  get: () => KernelStore
-) {
-  switch (message.type) {
-    case 'NEW_LOG':
-      get().addLog(message.payload);
-      break;
-
-    case 'NEW_RUNTIME_EVENT':
-      get().addRuntimeEvent(message.payload);
-      break;
-
-    case 'BATCH_LOGS':
-      get().addLogs(message.payload);
-      break;
-
-    case 'CLEAR_LOGS':
-      set({ logs: [], stats: { ...get().stats, totalLogs: 0, filteredCount: 0 } });
-      break;
-
-    case 'CLEAR_RUNTIME_EVENTS':
-      set({ runtimeEvents: [] });
-      break;
-
-    case 'CONNECTION_STATE':
-      set({ connection: message.payload });
-      break;
-
-    case 'RUNTIME_EVENT_CONNECTION_STATE':
-      set({ runtimeEventConnection: message.payload });
-      break;
-
-    case 'FILTER_UPDATE':
-      set({ filters: message.payload });
-      break;
-
-    case 'UI_UPDATE':
-      set({ ui: { ...get().ui, ...message.payload } });
-      break;
-
-    case 'REQUEST_SYNC':
-      // Primary window responds with full state
-      if (get()._isPrimaryWindow && get()._broadcastChannel) {
-        const state = get();
-        state._broadcastChannel!.postMessage({
-          type: 'FULL_SYNC',
-          payload: {
-            logs: state.logs,
-            runtimeEvents: state.runtimeEvents,
-            filters: state.filters,
-            ui: state.ui,
-          },
-        });
-      }
-      break;
-
-    case 'FULL_SYNC':
-      // Secondary window receives full state
-      if (!get()._isPrimaryWindow) {
-        set({
-          logs: message.payload.logs,
-          runtimeEvents: message.payload.runtimeEvents,
-          filters: message.payload.filters,
-          ui: message.payload.ui,
-        });
-      }
-      break;
-
-    case 'PRIMARY_HEARTBEAT':
-      // Update primary window timestamp
-      localStorage.setItem(PRIMARY_WINDOW_KEY, message.timestamp.toString());
-      break;
-  }
-}
-
-/**
- * Start primary window heartbeat
- */
-function startPrimaryHeartbeat(get: () => KernelStore) {
-  const heartbeat = setInterval(() => {
-    const state = get();
-    if (state._isPrimaryWindow) {
-      const now = Date.now();
-      localStorage.setItem(PRIMARY_WINDOW_KEY, now.toString());
-
-      // Broadcast heartbeat to other windows
-      if (state._broadcastChannel) {
-        state._broadcastChannel.postMessage({
-          type: 'PRIMARY_HEARTBEAT',
-          timestamp: now,
-        });
-      }
-    }
-  }, HEARTBEAT_INTERVAL);
-
-  return heartbeat;
-}
-
-/**
- * Create Zustand store
- */
-export const useKernelStore = create<KernelStore>()(
-  devtools(
-    persist(
-      (set, get) => ({
-        // Initial state
-        logs: [],
-        runtimeEvents: [],
-        traceGroups: new Map(),
-        connection: {
-          status: 'disconnected',
-          error: null,
-          connectedAt: null,
-          reconnectAttempts: 0,
-          lastPingTime: null,
-        },
-        runtimeEventConnection: {
-          status: 'disconnected',
-          error: null,
-          connectedAt: null,
-          lastEventId: null,
-        },
-        filters: {
-          logLevel: null,
-          loggerNamespace: '',
-          searchText: '',
-        },
-        ui: {
-          autoScroll: true,
-          isPaused: false,
-          maxBufferSize: 1000,
-        },
-        stats: {
-          totalLogs: 0,
-          filteredCount: 0,
-          connectionUptime: 0,
-          logsPerSecond: 0,
-        },
-
-        // Internal state
-        _ws: null,
-        _eventSource: null,
-        _broadcastChannel: null,
-        _isPrimaryWindow: false,
-        _reconnectTimer: null,
-        _statsUpdateTimer: null,
-        _manualDisconnecting: false,
-
-        // Computed selector
-        filteredLogs: () => {
-          const { logs, filters } = get();
-          return logs.filter((log) => {
-            // Level filter
-            if (filters.logLevel && log.level !== filters.logLevel) {
-              return false;
-            }
-
-            // Logger namespace filter
-            if (
-              filters.loggerNamespace &&
-              !log.logger.startsWith(filters.loggerNamespace)
-            ) {
-              return false;
-            }
-
-            // Search text filter
-            if (filters.searchText) {
-              const search = filters.searchText.toLowerCase();
-              return (
-                log.message.toLowerCase().includes(search) ||
-                log.logger.toLowerCase().includes(search)
-              );
-            }
-
-            return true;
-          });
-        },
-
-        // Connection actions
-        connect: () => {
-          const state = get();
-          if (
-            state._ws?.readyState === WebSocket.OPEN ||
-            state._ws?.readyState === WebSocket.CONNECTING
-          ) {
-            console.log('[KernelStore] Already connected');
-            return;
-          }
-
-          console.log('[KernelStore] Connecting...');
-          set({
-            connection: {
-              ...state.connection,
-              status: 'connecting',
-            },
-            _manualDisconnecting: false,
-          });
-
-          // Determine if this is the primary window
-          try {
-            const isPrimary = checkAndClaimPrimaryWindow();
-            set({ _isPrimaryWindow: isPrimary });
-
-            console.log(
-              `[KernelStore] Window role: ${isPrimary ? 'PRIMARY' : 'SECONDARY'}`
-            );
-
-            if (!state._broadcastChannel) {
-              const bc = new BroadcastChannel(BROADCAST_CHANNEL_NAME);
-              bc.onmessage = (event) => handleBroadcastMessage(event.data, set, get);
-              set({ _broadcastChannel: bc });
-              console.log('[KernelStore] BroadcastChannel initialized');
-            }
-
-            if (isPrimary) {
-              initializeWebSocket(set, get);
-              initializeRuntimeEventStream(set, get);
-              const heartbeatTimer = startPrimaryHeartbeat(get);
-              set({ _statsUpdateTimer: heartbeatTimer });
-            } else {
-              const bc = get()._broadcastChannel;
-              if (bc) {
-                bc.postMessage({ type: 'REQUEST_SYNC' });
-                console.log('[KernelStore] Requested full sync from primary');
-              }
-
-          set({
-            connection: {
-              ...state.connection,
-              status: 'connected',
-              connectedAt: Date.now(),
-            },
-            runtimeEventConnection: {
-              ...state.runtimeEventConnection,
-              status: 'connected',
-              connectedAt: Date.now(),
-            },
-          });
-            }
-          } catch (error) {
-            const message =
-              error instanceof Error ? error.message : 'Kernel connection initialization failed';
-            console.error('[KernelStore] Failed to start connection:', error);
-            set({
-              connection: {
-                ...get().connection,
-                status: 'error',
-                error: message,
-              },
-            });
-          }
-        },
-
-        disconnect: () => {
-          console.log('[KernelStore] Disconnecting...');
-          const state = get();
-          const hasWs = Boolean(state._ws);
-          if (hasWs) {
-            set({ _manualDisconnecting: true });
-          }
-
-          // Close WebSocket
-          if (state._ws) {
-            state._ws.close();
-            set({ _ws: null });
-          }
-
-          // Close BroadcastChannel
-          if (state._eventSource) {
-            state._eventSource.close();
-            set({ _eventSource: null });
-          }
-
-          // Close BroadcastChannel
-          if (state._broadcastChannel) {
-            state._broadcastChannel.close();
-            set({ _broadcastChannel: null });
-          }
-
-          // Clear timers
-          if (state._reconnectTimer) {
-            clearTimeout(state._reconnectTimer);
-            set({ _reconnectTimer: null });
-          }
-          if (state._statsUpdateTimer) {
-            clearInterval(state._statsUpdateTimer);
-            set({ _statsUpdateTimer: null });
-          }
-
-          // Clear primary window flag
-          if (state._isPrimaryWindow) {
-            localStorage.removeItem(PRIMARY_WINDOW_KEY);
-          }
-
-          set({
-            connection: {
-              status: 'disconnected',
-              error: null,
-              connectedAt: null,
-              reconnectAttempts: 0,
-              lastPingTime: null,
-            },
-            _isPrimaryWindow: false,
-            _manualDisconnecting: hasWs,
-            runtimeEventConnection: {
-              ...state.runtimeEventConnection,
-              status: 'disconnected',
-              error: null,
-              connectedAt: null,
-            },
-          });
-        },
-
-        reconnect: () => {
-          console.log('[KernelStore] Manual reconnect triggered');
-          get().disconnect();
-          setTimeout(() => get().connect(), 1000);
-        },
-
-        connectRuntimeEvents: () => {
-          initializeRuntimeEventStream(set, get);
-        },
-
-        disconnectRuntimeEvents: () => {
-          const state = get();
-          if (state._eventSource) {
-            state._eventSource.close();
-          }
-          set({
-            _eventSource: null,
-            runtimeEventConnection: {
-              ...state.runtimeEventConnection,
-              status: 'disconnected',
-              error: null,
-              connectedAt: null,
-            },
-          });
-        },
-
-        // Log actions
-        addLog: (log) => {
-          const state = get();
-          if (state.ui.isPaused) return;
-
-          const newLog: LogEntry = {
-            ...log,
-            id: crypto.randomUUID(),
-          };
-
-          set((state) => {
-            let newLogs = [...state.logs, newLog];
-
-            // Enforce buffer size limit
-            if (newLogs.length > state.ui.maxBufferSize) {
-              newLogs = newLogs.slice(-state.ui.maxBufferSize);
-            }
-
-            return {
-              logs: newLogs,
-              traceGroups: groupLogsByTrace(newLogs),
-              stats: {
-                ...state.stats,
-                totalLogs: state.stats.totalLogs + 1,
-              },
-            };
-          });
-        },
-
-        addLogs: (logs) => {
-          const state = get();
-          if (state.ui.isPaused) return;
-
-          const newLogs: LogEntry[] = logs.map((log) => ({
-            ...log,
-            id: crypto.randomUUID(),
-          }));
-
-          set((state) => {
-            let allLogs = [...state.logs, ...newLogs];
-
-            // Enforce buffer size limit
-            if (allLogs.length > state.ui.maxBufferSize) {
-              allLogs = allLogs.slice(-state.ui.maxBufferSize);
-            }
-
-            return {
-              logs: allLogs,
-              traceGroups: groupLogsByTrace(allLogs),
-              stats: {
-                ...state.stats,
-                totalLogs: state.stats.totalLogs + newLogs.length,
-              },
-            };
-          });
-        },
-
-        clearLogs: () => {
-          set({
-            logs: [],
-            traceGroups: new Map(),
-            stats: {
-              totalLogs: 0,
-              filteredCount: 0,
-              connectionUptime: get().stats.connectionUptime,
-              logsPerSecond: 0,
-            },
-          });
-
-          // Broadcast clear to other windows
-          const state = get();
-          if (state._isPrimaryWindow && state._broadcastChannel) {
-            state._broadcastChannel.postMessage({ type: 'CLEAR_LOGS' });
-          }
-        },
-
-        addRuntimeEvent: (event) => {
-          const state = get();
-          useMemoryTaskStore.getState().applyRuntimeEvent(event);
-          if (state.ui.isPaused) return;
-
-          set((state) => {
-            let runtimeEvents = [...state.runtimeEvents, event];
-            if (runtimeEvents.length > state.ui.maxBufferSize) {
-              runtimeEvents = runtimeEvents.slice(-state.ui.maxBufferSize);
-            }
-            return {
-              runtimeEvents,
-              runtimeEventConnection: {
-                ...state.runtimeEventConnection,
-                lastEventId: event.event_id,
-              },
-            };
-          });
-        },
-
-        clearRuntimeEvents: () => {
-          set({ runtimeEvents: [] });
-          const state = get();
-          if (state._isPrimaryWindow && state._broadcastChannel) {
-            state._broadcastChannel.postMessage({ type: 'CLEAR_RUNTIME_EVENTS' });
-          }
-        },
-
-        // Filter actions
-        setLogLevel: (level) => {
-          set((state) => ({
-            filters: { ...state.filters, logLevel: level },
-          }));
-        },
-
-        setLoggerNamespace: (namespace) => {
-          set((state) => ({
-            filters: { ...state.filters, loggerNamespace: namespace },
-          }));
-        },
-
-        setSearchText: (text) => {
-          set((state) => ({
-            filters: { ...state.filters, searchText: text },
-          }));
-        },
-
-        clearFilters: () => {
-          set({
-            filters: {
-              logLevel: null,
-              loggerNamespace: '',
-              searchText: '',
-            },
-          });
-        },
-
-        // UI actions
-        toggleAutoScroll: () => {
-          set((state) => ({
-            ui: { ...state.ui, autoScroll: !state.ui.autoScroll },
-          }));
-        },
-
-        togglePause: () => {
-          set((state) => ({
-            ui: { ...state.ui, isPaused: !state.ui.isPaused },
-          }));
-        },
-
-        setMaxBufferSize: (size) => {
-          set((state) => ({
-            ui: { ...state.ui, maxBufferSize: size },
-          }));
-        },
-
-        toggleTraceCollapse: (trace_id) => {
-          set((state) => {
-            const newGroups = new Map(state.traceGroups);
-            const trace = newGroups.get(trace_id);
-            if (trace) {
-              trace.collapsed = !trace.collapsed;
-            }
-            return { traceGroups: newGroups };
-          });
-        },
-
-        toggleSpanCollapse: (trace_id, span_name) => {
-          set((state) => {
-            const newGroups = new Map(state.traceGroups);
-            const trace = newGroups.get(trace_id);
-            if (trace) {
-              const span = trace.spans.get(span_name);
-              if (span) {
-                span.collapsed = !span.collapsed;
-              }
-            }
-            return { traceGroups: newGroups };
-          });
-        },
-      }),
-      {
-        name: 'kernel-store',
-        partialize: (state) => ({
-          // Only persist UI preferences
-          filters: state.filters,
-          ui: state.ui,
-        }),
-      }
-    ),
-    { name: 'KernelStore' }
-  )
-);
+export { useKernelStore } from './kernel';
