+++ frontend/src/stores/memory/index.ts
@@ -1,0 +1,2 @@
+export { useMemoryTaskStore } from './taskStore';
+export { useMemoryViewStore } from './viewStore';
