+++ frontend/src/stores/toastStore.ts
@@ -1,34 +1,2 @@
-import { create } from 'zustand';
-
-export type ToastType = 'success' | 'error' | 'info' | 'warning';
-
-export interface Toast {
-  id: string;
-  message: string;
-  type: ToastType;
-  duration?: number;
-}
-
-interface ToastState {
-  toasts: Toast[];
-  addToast: (message: string, type?: ToastType, duration?: number) => void;
-  removeToast: (id: string) => void;
-}
-
-export const useToastStore = create<ToastState>((set) => ({
-  toasts: [],
-  addToast: (message, type = 'info', duration = 3000) => {
-    const id = Math.random().toString(36).substring(2, 9);
-    const toast = { id, message, type, duration };
-    set((state) => ({ toasts: [...state.toasts, toast] }));
-    
-    if (duration > 0) {
-      setTimeout(() => {
-        set((state) => ({ toasts: state.toasts.filter((t) => t.id !== id) }));
-      }, duration);
-    }
-  },
-  removeToast: (id) => {
-    set((state) => ({ toasts: state.toasts.filter((t) => t.id !== id) }));
-  },
-}));
+export { useToastStore } from './ui';
+export type { Toast, ToastType } from './ui';
