+++ frontend/src/stores/topic/index.ts
@@ -1,0 +1,1 @@
+export { useTopicStore } from './topicStore';
