+++ frontend/src/stores/ui/index.ts
@@ -1,0 +1,3 @@
+export { useChatUiStore } from './appUiStore';
+export { useToastStore } from './toastStore';
+export type { Toast, ToastType } from './toastStore';
